Fix: allow for proper chaining of operations
The issue with this bug was that I wasn't checking whether an operation
was in progress when I updated the display from a number button. It now
checks to see if an operation is in progress and, if so, stores the
displayed value to memory instead of appending to it.